Release 2.0.4
        This release is a public release.

        By popular demand, the default log file directory is now PREFIX/var/log/yadifa.  It can be set using --with-logdir=/my/dir

        Improved build mechanism.
            It has been tested to work automatically on Linux, FreeBSD, OSX, SunOS.
            RedHat familly builds will use -O2 as maximum optimisations.

            Note that some optional features are now enabled by default but can be disabled.

        Fixes:
        - fixed an issue with the AXFR transfer where the serial number would not be properly taken into account
        - fixed an issue with the notify mechanism that could occur if the server was only listening to 127.0.0.1
        - fixed an issue with bogus DNSKEY records that may potentially lead to a crash in openssl
        - fixed a reported potential "tmpfile" vulnerability on DEBUG builds (generated with make debug)
        - fixed an issue with IPv6 connections on some architectures
        - typos fixes
        - minor fixes and improvements

        Architecture portability enhancements.

        On Solaris, if no --enable-force32bits nor --enable-force64bits is set, then 64 bits will be forced (fixes an issue at link-time)

        ELF 64-bit MSB executable SPARCV9 Version 1, UltraSPARC3 Extensions Required, dynamically linked, not stripped, no debugging information available

        PATH=/opt/csw/bin:/usr/ccs/bin:$PATH ./configure --enable-force32bits
        PATH=/opt/csw/bin:/usr/ccs/bin:$PATH make

        Architecture portability enhancements.

    FreeBSD 9
        FreeBSD dnode3 9.0-RELEASE-p3 FreeBSD 9.0-RELEASE-p3 #0: Tue Jun 12 02:52:29 UTC 2012     root@amd64-builder.daemonology.net:/usr/obj/usr/src/sys/GENERIC  amd64
        gcc (GCC) 4.2.1 20070831 patched [FreeBSD]
        ELF 64-bit LSB executable, x86-64, version 1 (FreeBSD), dynamically linked (uses shared libs), for FreeBSD 9.0 (900044), not stripped

    Ubuntu
        Linux dnode10 3.2.0-49-generic #75-Ubuntu SMP Tue Jun 18 17:39:32 UTC 2013 x86_64 x86_64 x86_64 GNU/Linux
        gcc (Ubuntu/Linaro 4.6.3-1ubuntu5) 4.6.3
        ELF 64-bit LSB executable, x86-64, version 1 (SYSV), dynamically linked (uses shared libs), for GNU/Linux 2.6.24, BuildID[sha1]=0xe3b8601b9b5e59f8c9ce519cacbe9b8ff544ff1d, not stripped

    OSX
        Darwin RD-Mac-Mini.local 13.3.0 Darwin Kernel Version 13.3.0: Tue Jun  3 21:27:35 PDT 2014; root:xnu-2422.110.17~1/RELEASE_X86_64 x86_64
        Apple LLVM version 5.1 (clang-503.0.40) (based on LLVM 3.4svn)
        Mach-O 64-bit executable x86_64

        Architecture portability enhancements.

        uname -a
        gcc --version
        file yadifad

    YellowDog Linux
        Linux 2.6.29-3.ydl61.4 #1 SMP Mon Sep 7 14:50:27 PDT 2009 ppc64 ppc64 ppc64 GNU/Linux
        gcc (GCC) 4.1.2 20080704 (Red Hat 4.1.2-44)
        ELF 32-bit MSB executable, PowerPC or cisco 4500, version 1 (SYSV), for GNU/Linux 2.6.9, dynamically linked (uses shared libs), for GNU/Linux 2.6.9, not stripped

        note: using --enable-force64bits failed because of ssl, no simple/quick way to install openssl-devel.ppc64 seemed available

    Debian PPC64
        Linux 3.2.0-3-powerpc64 #1 SMP Mon Jul 23 08:03:56 UTC 2012 ppc64 GNU/Linux
        gcc (Debian 4.6.3-8) 4.6.3
        ELF 32-bit MSB executable, PowerPC or cisco 4500, version 1 (SYSV), dynamically linked (uses shared libs), for GNU/Linux 2.6.26, BuildID[sha1]=0xedc47c984a4af7eb9a7ecbc0f135e4d064ba08f0, with unknown capability 0x41000000 = 0x13676e75, with unknown capability 0x10000 = 0xb0401, not stripped

        note: using --enable-force64bits failed because of ssl, no simple/quick way to install openssl-devel.ppc64 seemed available

+++ lib/dnsdb/include/dnsdb/zdb_alloc.h
@@ -94,8 +94,8 @@ extern "C" {
    
 #define ZDB_ALLOC_PG_SIZE_COUNT         256 // 2K
 #define ZDB_ALLOC_PG_PAGEABLE_MAXSIZE   (ZDB_ALLOC_PG_SIZE_COUNT * 8) /* x 8 because we are going by 8 increments */
-#define ZDB_MALLOC_SIZE_TO_PAGE(size_)  (((size_)-1)>>3)
-#define ZDB_ALLOC_CANHANDLE(size_)      ((size_)<=ZDB_ALLOC_PG_PAGEABLE_MAXSIZE)
+#define ZDB_MALLOC_SIZE_TO_PAGE(size_)  ((s32)(((size_)-1)>>3))
+#define ZDB_ALLOC_CANHANDLE(size_)      ((s32)((size_)<=ZDB_ALLOC_PG_PAGEABLE_MAXSIZE))
 
 // prepares the zdb_alloc tables
     
@@ -244,10 +244,10 @@ void zdb_mfree_unaligned(void* ptr);
 
 #define ZALLOC_ARRAY_RESIZE(type_,array_,count_,newcount_)			\
 {										\
-    int zalloc_new_count = (newcount_);						\
-    if((count_) != zalloc_new_count)						\
+    int zalloc_new_count = (u32)(newcount_);						\
+    if(((u32)(count_)) != zalloc_new_count)						\
     {										\
-	if( ZDB_MALLOC_SIZE_TO_PAGE(sizeof(type_)*(count_)) !=                  \
+	if( ZDB_MALLOC_SIZE_TO_PAGE(sizeof(type_)*((u32)(count_))) !=                  \
 	    ZDB_MALLOC_SIZE_TO_PAGE(sizeof(type_)*zalloc_new_count))		\
 	{									\
 	    type_* __tmp__;							\
@@ -256,14 +256,14 @@ void zdb_mfree_unaligned(void* ptr);
 	    {									\
 		ZALLOC_ARRAY_OR_DIE(type_*,__tmp__,sizeof(type_)*zalloc_new_count, \
 				    ZALLOC_ARRAY_RESIZE_TAG);			\
-		MEMCOPY(__tmp__,(array_),sizeof(type_)*MIN((count_),zalloc_new_count));	\
+		MEMCOPY(__tmp__,(array_),sizeof(type_)*MIN(((u32)(count_)),zalloc_new_count));	\
 	    }									\
 	    else								\
 	    {									\
 		__tmp__ = NULL;							\
 	    }									\
 										\
-	    ZFREE_ARRAY((array_),sizeof(type_)*(count_));			\
+	    ZFREE_ARRAY((array_),sizeof(type_)*((u32)(count_)));			\
 	    array_ = __tmp__;							\
 	    count_ = newcount_;							\
 	}									\
